+++ lib/prawnto/action_view_mixin.rb
@@ -1,0 +1,13 @@
+module Prawnto
+  module ActionViewMixin
+
+  private
+
+    def _prawnto_compile_setup
+      compile_support = CompileSupport.new(controller)
+      @prawnto_options = compile_support.options
+    end
+
+  end
+end
+
