--- lib/prawnto/action_view_mixins/action_view_mixin.rb
@@ -1,15 +1,0 @@
-module Prawnto
-  module ActionViewMixins
-    module ActionViewMixin
-
-    private
-
-      def _prawnto_compile_setup
-        compile_support = CompileSupport.new(controller)
-        @prawnto_options = compile_support.options
-      end
-
-    end
-  end
-end
-
