+++ spec/units/compile_support_spec.rb
@@ -1,6 +1,6 @@
 require File.expand_path("../spec_helper.rb", File.dirname(__FILE__))
 
-describe Prawnto::ActionViewMixins::CompileSupport do
+describe Prawnto::CompileSupport do
   
   before do
     @request = mock()
@@ -14,26 +14,26 @@ describe Prawnto::ActionViewMixins::CompileSupport do
 
   describe "#set_disposition" do
     before do
-      Prawnto::ActionViewMixins::CompileSupport.any_instance.stubs(:set_pragma).returns(true)
-      Prawnto::ActionViewMixins::CompileSupport.any_instance.stubs(:set_cache_control).returns(true)
-      Prawnto::ActionViewMixins::CompileSupport.any_instance.stubs(:set_content_type).returns(Mime::PDF)
+      Prawnto::CompileSupport.any_instance.stubs(:set_pragma).returns(true)
+      Prawnto::CompileSupport.any_instance.stubs(:set_cache_control).returns(true)
+      Prawnto::CompileSupport.any_instance.stubs(:set_content_type).returns(Mime::PDF)
     end
     
     it "default" do
       @headers.expects("[]=").with("Content-Disposition", "inline").once
-      Prawnto::ActionViewMixins::CompileSupport.new(@controller)
+      Prawnto::CompileSupport.new(@controller)
     end
 
     it "inline with filename" do
       @controller.stubs(:compute_prawnto_options).returns({:filename => "xxx.pdf", :inline => true})
       @headers.expects("[]=").with("Content-Disposition", "inline;filename=\"xxx.pdf\"").once
-      Prawnto::ActionViewMixins::CompileSupport.new(@controller)
+      Prawnto::CompileSupport.new(@controller)
     end
     
     it "attachment with filename" do
       @controller.stubs(:compute_prawnto_options).returns({:filename => "xxx.pdf", :inline => false})
       @headers.expects("[]=").with("Content-Disposition", "attachment;filename=\"xxx.pdf\"").once
-      Prawnto::ActionViewMixins::CompileSupport.new(@controller)
+      Prawnto::CompileSupport.new(@controller)
     end
 
   end
